docu_assist/requirements.txt
crewai
crewai[tools]
python-dotenv
arize-phoenix==3.24.0
qdrant-client